apiflask-agendacomercial-dev/requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
apispec==5.2.2
astroid==2.11.6
autopep8==1.5.7
build==0.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.5
cx-Oracle==8.3.0
dill==0.3.5.1
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.0.1
flask-apispec==0.11.0
#Flask-JWT==0.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-RESTful==0.3.9
flask-swagger==0.2.14
Flask-WTF==1.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
marshmallow==3.15.0
mccabe==0.7.0
numpy==1.24.2
openpyxl==3.1.2
packaging==21.3
pandas==1.5.3
platformdirs==2.5.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pydantic_core==2.6.3
PyJWT==2.4.0
pylint==2.14.3
pyparsing==3.0.9
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.26.0
six==1.16.0
SQLAlchemy==2.0.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typing_extensions==4.2.0
urllib3==1.26.6
uvicorn==0.23.2
waitress==2.0.0
webargs==8.1.0
Werkzeug==2.0.1
wrapt==1.14.1
WTForms==3.0.1
xmltodict==0.12.0
xmltojson==2.0.2
yarg==0.1.9
